requirements.txt
arrow==4.0.1
certifi==2020.11.8
cfbd==3.3.3
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
pyreadr==0.3.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
